\name{CellID-package}
\alias{CellID-package}
\alias{CellID}
\docType{package}
\title{Multiple Correspondence Analysis on Single Cell for Joint Dimensionality Reduction of Gene and Cell, Cells Geneset Extraction and Geneset Enrichment Analysis}
\description{
\packageDescription{CellID}
}
\author{
\strong{Maintainer}: Akira Cortal \email{akira.cortal@institutimagine.org}

Authors:
\itemize{
\item Akira Cortal
\item Antonio Rausell
}
}
\references{
\itemize{
\item{
Rausell, A., Juan, D., Pazos, F., & Valencia, A. (2010). Protein interactions and ligand binding: from protein subfamilies to functional specificity. Proceedings of the National Academy of Sciences of the United States of America, 107(5), 1995–2000. \url{https://doi.org/10.1073/pnas.0908044107}
}

\item
Aan, Z., & Greenacre, M. (2011). Biplots of fuzzy coded data. Fuzzy Sets and Systems, 183(1), 57–71. \url{https://doi.org/10.1016/j.fss.2011.03.007}

\item
Alexey Sergushichev. An algorithm for fast preranked gene set enrichment analysis using cumulative statistic calculation. bioRxiv (2016), 
\url{https://doi.org/10.1101/060012 }

\item
Stuart and Butler et al. Comprehensive integration of single cell data. bioRxiv (2018).    \url{https://doi.org/10.1101/460147}

\item
Aaron Lun and Davide Risso (2019). SingleCellExperiment: S4 Classes for Single Cell Data. R package version 1.4.1.
}
}
\keyword{SCRNASeq, MCA, GSEA, Gene Sets}
\seealso{
\itemize{
\item
McCarthy, D. J., Campbell, K. R., Lun, A. T. L., & Wills, Q. F. (2017). Scater: pre-processing, quality control, normalization and visualization of single-cell RNA-seq data in R. Bioinformatics, 33(8), btw777. \url{https://doi.org/10.1093/bioinformatics/btw777}

\item
Amezquita, R. A., Carey, V. J., Carpp, L. N., Geistlinger, L., Lun, A. T. L., Marini, F., … Hicks, S. C. (2019). Orchestrating Single-Cell Analysis with Bioconductor. BioRxiv, 590562. \url{https://doi.org/10.1101/590562}
}
}