% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{DistSort}
\alias{DistSort}
\title{Sort Gene Cell Distance Matrix}
\usage{
DistSort(distance)
}
\arguments{
\item{distance}{distance matrix with features at rows and cell at columns}
}
\value{
list of ranking of genes by cells
}
\description{
Sort Gene Cell Distance Matrix
}
