% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcaAndReduction.R
\name{GetGeneCellCoordinates}
\alias{GetGeneCellCoordinates}
\title{GeneCellCoordinates}
\usage{
GetGeneCellCoordinates(X, reduction, dims, features)
}
\arguments{
\item{X}{Seurat or SingleCellExperiment Object}

\item{reduction}{Which dimensionality reduction to use, must be based on MCA.}

\item{dims}{A vector of integers indicating which dimensions to use with reduction embeddings and loadings for distance calculation.}

\item{features}{Character vector of feature names to subset feature coordinates. If not specified will take all features available from specified reduction Loadings.}
}
\value{
A matrix with gene and cell coordinates of MCA
}
\description{
Get coordinates of both cells and features in a matrix
}
