% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_biological}
\alias{GO_biological}
\title{Biological Process GO Terms}
\format{A list of 3857 gene onthology terms with the corresponding genes.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp#C5}
}
\usage{
GO_biological
}
\description{
A list of gene vectors representing the different biological process as discribed in Gene Onthology.
}
\references{
The Gene Ontology project in 2008, The Gene Ontology Consortium Nucleic Acids Research, Volume 36, Issue suppl_1, January 2008, Pages D440–D444,
}
\keyword{datasets}
