% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hallmark}
\alias{Hallmark}
\title{Hallmark Pathways from MSigDB}
\format{
A named list of length 50 containing Hallmark gene sets.
}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/download_file.jsp?filePath=/resources/msigdb/6.2/h.all.v6.2.symbols.gmt}
}
\usage{
Hallmark
}
\description{
A dataset containing the Hallmark gene sets from MSigDB.
}
\references{
Liberzon A, Birger C, Thorvaldsdóttir H, Ghandi M, Mesirov JP, Tamayo P. The Molecular Signatures Database (MSigDB) hallmark gene set collection. Cell Syst. 2015 Dec 23;1(6):417-425.
}
\keyword{datasets}
