% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HgProteinCodingGenes}
\alias{HgProteinCodingGenes}
\title{Homo Sapiens Protein Coding Genes}
\format{
A list of 19308 gene onthology terms with the corresponding genes.
}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp#C5}
}
\usage{
HgProteinCodingGenes
}
\description{
A gene list of human protein coding genes extracted from biomaRt.
}
\references{
The Gene Ontology project in 2008, The Gene Ontology Consortium Nucleic Acids Research, Volume 36, Issue suppl_1, January 2008, Pages D440–D444,
}
\keyword{datasets}
