% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{integrateGSEASeurat}
\alias{integrateGSEASeurat}
\title{GSEA Results Integration in Seurat object}
\usage{
integrateGSEASeurat(X, gseaResults, assay.name)
}
\arguments{
\item{X}{a Seurat object}

\item{gseaResults}{Results of GSEA obtained by RunCellGSEA or RunGroupGSEA}

\item{assay.name}{The name of the new Seurat assay slot to include GSEA results}
}
\value{
a Seurat object with a new assay slot filled with GSE analysis resultsq
}
\description{
GSEA Results Integration in Seurat object
}
