% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{pairDist}
\alias{pairDist}
\title{Distance Calculation}
\usage{
pairDist(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}
}
\value{
A Distance Matrix
}
\description{
Small function to calculate quickly the distance between rows of two matrix.
}
