% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotReducedDimMC}
\alias{plotReducedDimMC}
\title{Scater plotReducedDim for MCA like dimensionality Reduction}
\usage{
plotReducedDimMC(
  X,
  reduction = "MCA",
  dims = c(1, 2),
  features = NULL,
  size.feature = 3,
  size.feature.text = 5,
  as.text = FALSE,
  ...
)
}
\arguments{
\item{X}{a Single Cell Experiment Object}

\item{reduction}{Which dimensionality reduction to use. If not specified, searches for mca.}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{features}{character vector of features to plot, must be present in the specified dimension loadings}

\item{size.feature}{integer indicating size of geom_point for features}

\item{size.feature.text}{integer indicating size of geom_text for features}

\item{as.text}{logical indicating as to include text label for feature plotting, will produce warning if TRUE and length(features) > 50.}

\item{...}{Other arguments passed to plotReducedDim}
}
\value{
A ggplot object
}
\description{
Small modification of the Scater plotReducedDim function to enable plotting features for mca like dimensionality reduction.
}
\examples{
scePBMC <- as.SingleCellExperiment(seuratPbmc)
scePBMC <- RunMCA(scePBMC, nmcs = 5)
plotReducedDimMC(scePBMC)
}
