% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcaAndReduction.R
\name{RunMCTSNE}
\alias{RunMCTSNE}
\alias{RunMCTSNE.Seurat}
\alias{RunMCTSNE.SingleCellExperiment}
\title{Run TSNE on MCA fetures and cells coordinates}
\usage{
RunMCTSNE(X, reduction, dims, features, reduction.name, ...)

\method{RunMCTSNE}{Seurat}(
  X,
  reduction = "mca",
  dims = seq(50),
  features = NULL,
  reduction.name = "mctsne",
  assay = DefaultAssay(X),
  ...
)

\method{RunMCTSNE}{SingleCellExperiment}(
  X,
  reduction = "MCA",
  dims = seq(50),
  features = NULL,
  reduction.name = "MCTSNE",
  ...
)
}
\arguments{
\item{X}{Seurat or SingleCellExperiment object}

\item{reduction}{Which dimensionality reduction to use, must be based on MCA.}

\item{dims}{A vector of integers indicating which dimensions to use with reduction embeddings and loadings for distance calculation.}

\item{features}{Character vector of feature names to subset feature coordinates. If not specified will take all features available from specified reduction Loadings.}

\item{reduction.name}{name of the created dimensionlaity reduction, default set to "mca" for Seurat and "MCA" for SCE.}

\item{...}{other arguments passed to methods or Rtsne::Rtsne}

\item{assay}{Seurat assay slot. When not specified set with DefaultAssay(X)}
}
\value{
Seurat or SingleCellExperiment object with MCTSNE stored in the reduction slot
}
\description{
Run TSNE on MCA fetures and cells coordinates
}
\examples{
seuratPbmc <- RunMCA(seuratPbmc, nmcs = 5)
seuratPbmc <- RunMCTSNE(seuratPbmc, dims = seq(5))
}
