% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seuratPbmc}
\alias{seuratPbmc}
\title{Seurat object of 400 PBMC cells}
\format{
A seurat object.
}
\source{
\url{https://s3-us-west-2.amazonaws.com/10x.files/samples/cell/pbmc3k/pbmc3k_filtered_gene_bc_matrices.tar.gz}
}
\usage{
seuratPbmc
}
\description{
A subset of the PBMC3k data from Seurat vignette. Normalisation, VariableFeatures, ScaleData and PCA has alreay been computed with default Seurat parameter.
}
\references{
Butler et al., Nature Biotechnology 2018.
}
\keyword{datasets}
