% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{pairDist}
\alias{pairDist}
\title{Distance Calculation with rdist}
\usage{
pairDist(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}
}
\value{
A Distance Matrix
}
\description{
Small modification of rdist from fields package to include column and rownames
}
