% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca.R
\name{RunMCA}
\alias{RunMCA}
\alias{RunMCA.matrix}
\alias{RunMCA.Seurat}
\alias{RunMCA.SingleCellExperiment}
\title{Run Multiple Correspondence Analysis}
\usage{
RunMCA(X, nmcs, features, reduction.name, slot, ...)

\method{RunMCA}{matrix}(X, nmcs = 50, features = NULL, reduction.name = "MCA", ...)

\method{RunMCA}{Seurat}(
  X,
  nmcs = 50,
  features = NULL,
  reduction.name = "mca",
  slot = "data",
  assay = DefaultAssay(X),
  ...
)

\method{RunMCA}{SingleCellExperiment}(
  X,
  nmcs = 50,
  features = NULL,
  reduction.name = "MCA",
  slot = "logcounts",
  ...
)
}
\arguments{
\item{X}{Seurat, SingleCellExperiment or matrix object}

\item{nmcs}{number of components to compute and store, default set to 30}

\item{features}{character vector of feature names. If not specified all features will be taken.}

\item{reduction.name}{name of the reduction default set to 'MCA' for SingleCellExperiment and mca}

\item{slot}{Which slot to pull expression data from? Default to logcounts for SingleCellExperiment and data for Seurat.}

\item{...}{other aruments passed to methods}

\item{assay}{Name of Assay MCA is being run on}
}
\value{
Seurat or SCE object with MCA calculation stored in the reductions slot.
}
\description{
Run Multiple Correspondence Analysis
}
\examples{
seuratPbmc <- RunMCA(seuratPbmc, nmcs = 5)
}
